projects/LHCbPR2BE/requirements.txt
Django==1.9.2
PyYAML==3.11
django-dotenv==1.4.1
django-extensions==1.5.3
django-filter==0.9.2
django-fixture-magic==0.0.7
git+https://github.com/mazurov/django-rest-swagger.git#egg=django-rest-swagger
djangorestframework==3.3.2
djangorestframework-jsonp==1.0.2
drf-extensions==0.2.7
flup==1.0.2
mysqlclient==1.3.6
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
wsgiref==0.1.2
